requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.7.0
dnspython==1.16.0
idna==2.9
motor==2.1.0
multidict==4.7.5
pymongo==3.10.1
python-dotenv==0.13.0
requests==2.23.0
requests-cache==0.5.2
urllib3==1.25.9
websockets==8.1
yarl==1.4.2